+++ src/components/link/link.pw.tsx
@@ -64,6 +64,105 @@ test.describe("check props for Link component", () => {
     await expect(linkElement).toHaveCSS("cursor", "not-allowed");
   });
 
+  test("when `underline` prop is 'always' and component is not hovered, it should apply text-decoration underline", async ({
+    mount,
+    page,
+  }) => {
+    await mount(
+      <LinkComponent underline="always" href="#">
+        Test Content
+      </LinkComponent>,
+    );
+    const linkElement = linkChildren(page);
+    await expect(linkElement).toHaveCSS(
+      "text-decoration",
+      "underline solid rgb(0, 126, 69)",
+    );
+  });
+
+  test("when `underline` prop is 'always' and component is hovered over, it should maintain text-decoration underline", async ({
+    mount,
+    page,
+  }) => {
+    await mount(
+      <LinkComponent underline="always" href="#">
+        Test Content
+      </LinkComponent>,
+    );
+    const linkElement = linkChildren(page);
+    await linkElement.hover();
+    await expect(linkElement).toHaveCSS(
+      "text-decoration",
+      "underline solid rgb(0, 103, 56)",
+    );
+  });
+
+  test("when `underline` prop is 'hover' and component is not hovered over, it should apply no text-decoration", async ({
+    mount,
+    page,
+  }) => {
+    await mount(
+      <LinkComponent underline="hover" href="#">
+        Test Content
+      </LinkComponent>,
+    );
+    const linkElement = linkChildren(page);
+    await expect(linkElement).toHaveCSS(
+      "text-decoration",
+      "none solid rgb(0, 126, 69)",
+    );
+  });
+
+  test("when `underline` prop is 'hover' and component is hovered over, it should apply the text-decoration underline", async ({
+    mount,
+    page,
+  }) => {
+    await mount(
+      <LinkComponent underline="hover" href="#">
+        Test Content
+      </LinkComponent>,
+    );
+    const linkElement = linkChildren(page);
+    await linkElement.hover();
+    await expect(linkElement).toHaveCSS(
+      "text-decoration",
+      "underline solid rgb(0, 103, 56)",
+    );
+  });
+
+  test("when `underline` prop is 'never' and component is not hovered over, it should apply no text-decoration", async ({
+    mount,
+    page,
+  }) => {
+    await mount(
+      <LinkComponent underline="never" href="#">
+        Test Content
+      </LinkComponent>,
+    );
+    const linkElement = linkChildren(page);
+    await expect(linkElement).toHaveCSS(
+      "text-decoration",
+      "none solid rgb(0, 126, 69)",
+    );
+  });
+
+  test("when `underline` prop is 'never' and component is hovered over, it should maintain no text-decoration", async ({
+    mount,
+    page,
+  }) => {
+    await mount(
+      <LinkComponent underline="never" href="#">
+        Test Content
+      </LinkComponent>,
+    );
+    const linkElement = linkChildren(page);
+    await linkElement.hover();
+    await expect(linkElement).toHaveCSS(
+      "text-decoration",
+      "none solid rgb(0, 103, 56)",
+    );
+  });
+
   test("should render with icon prop", async ({ mount, page }) => {
     await mount(<LinkComponent icon="add" />);
 
@@ -241,6 +340,27 @@ test.describe("check props for Link component", () => {
 
   (
     [
+      ["default", "rgb(77, 167, 126)"],
+      ["negative", "rgb(219, 115, 128)"],
+      ["neutral", "rgb(230, 235, 237)"],
+      ["subtle", "rgb(255, 255, 255)"],
+    ] as [LinkProps["variant"], string][]
+  ).forEach(([variant, defaultColor]) => {
+    test(`should render with variant prop set to ${variant} and isDarkBackground`, async ({
+      mount,
+      page,
+    }) => {
+      await mount(
+        <LinkComponentWithDarkBackground variant={variant} isDarkBackground />,
+      );
+
+      const linkElement = linkChildren(page);
+      await expect(linkElement).toHaveCSS("color", defaultColor);
+    });
+  });
+
+  (
+    [
       ["default", "rgb(0, 103, 56)"],
       ["negative", "rgb(162, 44, 59)"],
       ["neutral", "rgb(0, 103, 56)"],
@@ -263,13 +383,16 @@ test.describe("check props for Link component", () => {
       ["default", "rgb(25, 142, 89)"],
       ["negative", "rgb(208, 75, 92)"],
       ["neutral", "rgb(25, 142, 89)"],
+      ["subtle", "rgb(255, 255, 255)"],
     ] as [LinkProps["variant"], string][]
   ).forEach(([variant, hoverColor]) => {
     test(`should render with correct hover state with isDarkBackground prop set with ${variant} variant`, async ({
       mount,
       page,
     }) => {
-      await mount(<LinkComponent variant={variant} isDarkBackground />);
+      await mount(
+        <LinkComponentWithDarkBackground variant={variant} isDarkBackground />,
+      );
 
       const linkElement = linkChildren(page);
       await linkElement.hover();
@@ -455,6 +578,24 @@ test.describe("should check accessibility for Link component", () => {
         page,
       }) => {
         await mount(<LinkComponent variant={variant} />);
+
+        await checkAccessibility(page);
+      });
+    },
+  );
+
+  (["default", "negative", "neutral"] as LinkProps["variant"][]).forEach(
+    (variant) => {
+      test(`should pass accessibility tests with variant prop set to ${variant} and isDarkBackground`, async ({
+        mount,
+        page,
+      }) => {
+        await mount(
+          <LinkComponentWithDarkBackground
+            variant={variant}
+            isDarkBackground
+          />,
+        );
 
         await checkAccessibility(page);
       });