+++ src/components/text-editor/text-editor-test.stories.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import { Meta, StoryObj } from "@storybook/react";
 import { action } from "@storybook/addon-actions";
 
@@ -10,6 +10,7 @@ import Textbox from "../textbox";
 
 import useDebounce from "../../hooks/__internal__/useDebounce";
 import ReadOnlyEditor from "./__internal__";
+import createGuid from "../../__internal__/utils/helpers/guid";
 
 const meta: Meta<typeof TextEditor> = {
   title: "Text Editor/Test",
@@ -125,7 +126,7 @@ export const Functions = ({ ...props }: Partial<TextEditorProps>) => {
         onChange={handleChange}
         onLinkAdded={action("onLinkAdded")}
         onSave={action("onSave")}
-        value={defaultValue}
+        initialValue={defaultValue}
       />
     </Box>
   );
@@ -156,10 +157,10 @@ export const ReadOnlyEditorForNotes = () => {
       </Typography>
       <Box p={1} display="flex" gap={2} flexDirection="column">
         <Box p={1} backgroundColor="lightgray">
-          <ReadOnlyEditor value={defaultValue} />
+          <ReadOnlyEditor initialValue={defaultValue} />
         </Box>
         <Box p={1} backgroundColor="lightgray">
-          <ReadOnlyEditor value={htmlValue} />
+          <ReadOnlyEditor initialValue={htmlValue} />
         </Box>
       </Box>
     </Box>
@@ -188,6 +189,7 @@ export const OnChangeFormattedValues: Story = () => {
         onChange={(_, { htmlString, json }) => {
           setValueJSON(JSON.stringify(json, null, 2));
           setValueHTML(htmlString);
+          action("onChange")({ htmlString, json });
         }}
       />
       <div>
@@ -206,5 +208,33 @@ export const OnChangeFormattedValues: Story = () => {
 };
 OnChangeFormattedValues.storyName = "Change Handler With Formatted Values";
 OnChangeFormattedValues.parameters = {
+  chromatic: { disableSnapshot: true },
+};
+
+export const ExternalOverwrite: Story = () => {
+  const [, setValue] = useState("");
+  const [resetKey, setResetKey] = useState(createGuid);
+
+  useEffect(() => {
+    setTimeout(() => {
+      setValue(createFromHTML("<p>Message Changed</p>"));
+      setResetKey(createGuid()); // Reset editor by changing its key
+    }, 3000);
+  }, []);
+
+  return (
+    <TextEditor
+      key={resetKey}
+      labelText="Message"
+      initialValue={createFromHTML("<p>Hello world</p>")}
+      onChange={(value, formattedValues) => {
+        action("onChange")({ value, formattedValues });
+        setValue(createFromHTML(value));
+      }}
+    />
+  );
+};
+ExternalOverwrite.storyName = "Externally overwrite editor content";
+ExternalOverwrite.parameters = {
   chromatic: { disableSnapshot: true },
 };